[refactor] Create component folder and move Card.tsx
